+++ client/branded/src/search-ui/results/sidebar/SearchReference.tsx
@@ -229,30 +229,6 @@ To use this filter, the search query must contain \`type:diff\` or \`type:commit
         showSuggestions: false,
     },
     {
-        ...createQueryExampleFromString('has.tag({any string})'),
-        field: FilterType.repo,
-        description:
-            'DEPRECATED: Use "has.meta({tag}:)" instead. Search inside repositories that are tagged with the provided string.',
-        examples: ['repo:has.tag(ocaml)', '-repo:has.tag(golang)'],
-        showSuggestions: false,
-    },
-    {
-        ...createQueryExampleFromString('has({key:value})'),
-        field: FilterType.repo,
-        description:
-            'DEPRECATED: Use "has.meta({key}:{value})" instead. Search inside repositories associated with a key:value pair that matches the provided key:value pair.',
-        examples: ['repo:has(owner:jordan)', '-repo:has(team:search)'],
-        showSuggestions: false,
-    },
-    {
-        ...createQueryExampleFromString('has.key({any string})'),
-        field: FilterType.repo,
-        description:
-            'DEPRECATED: Use "has.meta({key})" instead. Search inside repositories that are associated with the given key, regardless of its value.',
-        examples: ['repo:has.key(owner)', '-repo:has.key(wip)'],
-        showSuggestions: false,
-    },
-    {
         ...createQueryExampleFromString('has.meta({key:value})'),
         field: FilterType.repo,
         description:
@@ -549,7 +525,7 @@ export interface SearchReferenceProps extends TelemetryProps, Pick<SearchQuerySt
     filter: string
 }
 
-const SearchReference = React.memo((props: SearchReferenceProps): ReactElement => {
+const SearchReference = React.memo(function SearchReference(props: SearchReferenceProps) {
     const [persistedTabIndex, setPersistedTabIndex] = useLocalStorage(SEARCH_REFERENCE_TAB_KEY, 0)
 
     const { setQueryState, telemetryService } = props
@@ -649,5 +625,7 @@ const SearchReference = React.memo((props: SearchReferenceProps): ReactElement =
 export function getSearchReferenceFactory(
     props: Omit<SearchReferenceProps, 'filter'>
 ): (filter: string) => React.ReactNode {
-    return (filter: string) => <SearchReference {...props} filter={filter} />
+    return function SearchReferenceFactory(filter: string) {
+        return <SearchReference {...props} filter={filter} />
+    }
 }