+++ client/web/src/storm/pages/LayoutPage/LayoutPage.tsx
@@ -31,7 +31,7 @@ import { SearchQueryStateObserver } from '../../../SearchQueryStateObserver'
 
 import styles from './LayoutPage.module.scss'
 
-const LazySetupWizard = lazyComponent(() => import('../../../setup-wizard'), 'SetupWizard')
+const LazySetupWizard = lazyComponent(() => import('../../../setup-wizard/SetupWizard'), 'SetupWizard')
 
 export interface LegacyLayoutProps extends LegacyLayoutRouteContext {
     children?: never
@@ -193,7 +193,7 @@ export const Layout: React.FC<LegacyLayoutProps> = props => {
                 />
             )}
 
-            <GlobalAlerts authenticatedUser={props.authenticatedUser} />
+            <GlobalAlerts authenticatedUser={props.authenticatedUser} isSourcegraphApp={props.isSourcegraphApp} />
             {!isSiteInit && !isSignInOrUp && !props.isSourcegraphDotCom && !disableFeedbackSurvey && (
                 <SurveyToast authenticatedUser={props.authenticatedUser} />
             )}